+++ Analytics-CSharp/Segment/Analytics/Utilities/EventPipeline.cs
@@ -221,7 +221,7 @@ namespace Segment.Analytics.Utilities
                         else
                         {
                             Analytics.Logger.Log(LogLevel.Error, message: "Error " + statusCode + " uploading " + url);
-                            shouldCleanup = _retryStateMachine.ShouldDeleteBatch(statusCode, retryAfterSeconds);
+                            shouldCleanup = retryStateMachine.ShouldDeleteBatch(statusCode, retryAfterSeconds);
                             if (shouldCleanup)
                             {
                                 _analytics.ReportInternalError(AnalyticsErrorType.NetworkServerRejected,